GraphQL Post edit and delete authorization [29/05/2023] modify delete_post.rb file for authorized post delete modfiy update_post.rb file for authorized post update

+++ app/graphql/mutations/posts/delete_post.rb
@@ -5,16 +5,19 @@ module Mutations
 
     # TODO: define arguments
     argument :id, ID, required: true
-
+    argument :user_id, Integer, required: true
     # TODO: define resolve method
-    def resolve(id:)
-      post = Post.find(id)
-      post.destroy
-      {
-        post: post
-      }
-    rescue ActiveRecord::RecordNotFound => e 
-      raise GraphQL::ExecutionError.new(e.message)
+    def resolve(id:,user_id:)
+      post = Post.find_by(id: id)
+      
+      raise GraphQL::ExecutionError.new("Record Not Found") if post.nil?
+      if post.user_id == user_id
+        post.destroy
+        { post: post }
+      else
+        { post: nil }
+        raise GraphQL::ExecutionError.new("You are not authorized to delete this post")
+      end
     end
   end
 end
